pyproject.toml
[tool.poetry]
name = "sauron-rule-engine"
version = "1.0.3"
description = "A simple rule engine implemented in python"
authors = ["jlugao <joaolhullier@gmail.com>"]
readme = "README.md"
repository = "https://github.com/jlugao/sauron-rule-engine"
license = "MIT"
packages = [
  { include = "sauron" },
  ]


[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.6.2"
"ruamel.yaml" = "^0.16.5"
blinker = "^1.4"
pylint = "^2.3"
mkdocs-material = "^7.1.11"
zuban = "^0.1.0"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^6.0"
coveralls = "<4.0"
ruff = "^0.1.0"
pre-commit = {version = "*",allow-prereleases = true}
pytest-cov = "^3.0.0"
ipython = "^7.16"
pylint = "^2.3"
rope = "^0.14.0"
mkdocs = "^1.2.3"
markdown-include = "^0.5.1"
pytest-azurepipelines = "^0.8.0"

[tool.ruff]
line-length = 79
target-version = "py38"
exclude = [
    ".git",
    ".hg",
    ".mypy_cache",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
]

[tool.ruff.lint]
select = ["E", "F", "W", "B", "C", "I"]
ignore = ["E203", "E266", "E501"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

